Added in saving and loading
Added in the basic saving and loading of menu's using Java's builtin serialization. Also edited the .gitignore to not include .ser files. Lastly, I put in a test to test these methods but it is a very user involved method, not your typical unit tests. These will have to be changed later to something better.